refactor: Extract panels to separate components
