+++ packages/form-builder/src/components/sidebar/panels/block-list-tree.js
@@ -1,0 +1,14 @@
+import {PanelHeader} from "@wordpress/components";
+import {__} from "@wordpress/i18n";
+import {__experimentalListView as ListView} from "@wordpress/block-editor";
+
+const BlockListTree = () => {
+    return (
+        <>
+            <PanelHeader label={__('List View', 'give')} />
+            <ListView showNestedBlocks={true} expandNested={true} />
+        </>
+    );
+};
+
+export default BlockListTree;
